+++ selfupdate/detect_test.go
@@ -2,7 +2,6 @@ package selfupdate
 
 import (
 	"fmt"
-	"os"
 	"regexp"
 	"strings"
 	"testing"
@@ -172,7 +171,7 @@ func TestInvalidSlug(t *testing.T) {
 		if err == nil {
 			t.Error(slug, "should be invalid slug")
 		}
-		if !strings.Contains(err.Error(), "Invalid slug format") {
+		if !strings.Contains(err.Error(), "invalid slug format") {
 			t.Error("Unexpected error for", slug, ":", err)
 		}
 	}
@@ -195,51 +194,6 @@ func TestNoReleaseFound(t *testing.T) {
 	}
 	if ok {
 		t.Fatal("Repo having no release should not be found")
-	}
-}
-
-func TestDetectFromBrokenGitHubEnterpriseURL(t *testing.T) {
-	up, err := NewUpdater(Config{APIToken: "hogehoge", EnterpriseBaseURL: "https://example.com"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	_, ok, _ := up.DetectLatest("foo/bar")
-	if ok {
-		t.Fatal("Invalid GitHub Enterprise base URL should raise an error")
-	}
-}
-
-func TestDetectFromGitHubEnterpriseRepo(t *testing.T) {
-	token := os.Getenv("GITHUB_ENTERPRISE_TOKEN")
-	base := os.Getenv("GITHUB_ENTERPRISE_BASE_URL")
-	repo := os.Getenv("GITHUB_ENTERPRISE_REPO")
-	if token == "" {
-		t.Skip("because token for GHE is not found")
-	}
-	if base == "" {
-		t.Skip("because base URL for GHE is not found")
-	}
-	if repo == "" {
-		t.Skip("because repo slug for GHE is not found")
-	}
-
-	up, err := NewUpdater(Config{APIToken: token, EnterpriseBaseURL: base})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	r, ok, err := up.DetectLatest(repo)
-	if err != nil {
-		t.Fatal("Fetch failed:", err)
-	}
-	if !ok {
-		t.Fatal(repo, "not found")
-	}
-	if r == nil {
-		t.Fatal("Release not detected")
-	}
-	if !r.Version.Equals(semver.MustParse("1.2.3")) {
-		t.Error("")
 	}
 }