+++ selfupdate/uncompress.go
@@ -6,12 +6,12 @@ import (
 	"bytes"
 	"compress/gzip"
 	"fmt"
-	"github.com/ulikunitz/xz"
 	"io"
-	"io/ioutil"
 	"path/filepath"
 	"runtime"
 	"strings"
+
+	"github.com/ulikunitz/xz"
 )
 
 func matchExecutableName(cmd, target string) bool {
@@ -44,7 +44,7 @@ func unarchiveTar(src io.Reader, url, cmd string) (io.Reader, error) {
 			break
 		}
 		if err != nil {
-			return nil, fmt.Errorf("Failed to unarchive .tar file: %s", err)
+			return nil, fmt.Errorf("failed to unarchive .tar file: %s", err)
 		}
 		_, name := filepath.Split(h.Name)
 		if matchExecutableName(cmd, name) {
@@ -53,7 +53,7 @@ func unarchiveTar(src io.Reader, url, cmd string) (io.Reader, error) {
 		}
 	}
 
-	return nil, fmt.Errorf("File '%s' for the command is not found in %s", cmd, url)
+	return nil, fmt.Errorf("file '%s' for the command is not found in %s", cmd, url)
 }
 
 // UncompressCommand uncompresses the given source. Archive and compression format is
@@ -66,15 +66,15 @@ func UncompressCommand(src io.Reader, url, cmd string) (io.Reader, error) {
 
 		// Zip format requires its file size for uncompressing.
 		// So we need to read the HTTP response into a buffer at first.
-		buf, err := ioutil.ReadAll(src)
+		buf, err := io.ReadAll(src)
 		if err != nil {
-			return nil, fmt.Errorf("Failed to create buffer for zip file: %s", err)
+			return nil, fmt.Errorf("failed to create buffer for zip file: %s", err)
 		}
 
 		r := bytes.NewReader(buf)
 		z, err := zip.NewReader(r, r.Size())
 		if err != nil {
-			return nil, fmt.Errorf("Failed to uncompress zip file: %s", err)
+			return nil, fmt.Errorf("failed to uncompress zip file: %s", err)
 		}
 
 		for _, file := range z.File {
@@ -85,13 +85,13 @@ func UncompressCommand(src io.Reader, url, cmd string) (io.Reader, error) {
 			}
 		}
 
-		return nil, fmt.Errorf("File '%s' for the command is not found in %s", cmd, url)
+		return nil, fmt.Errorf("file '%s' for the command is not found in %s", cmd, url)
 	} else if strings.HasSuffix(url, ".tar.gz") || strings.HasSuffix(url, ".tgz") {
 		log.Println("Uncompressing tar.gz file", url)
 
 		gz, err := gzip.NewReader(src)
 		if err != nil {
-			return nil, fmt.Errorf("Failed to uncompress .tar.gz file: %s", err)
+			return nil, fmt.Errorf("failed to uncompress .tar.gz file: %s", err)
 		}
 
 		return unarchiveTar(gz, url, cmd)
@@ -100,12 +100,12 @@ func UncompressCommand(src io.Reader, url, cmd string) (io.Reader, error) {
 
 		r, err := gzip.NewReader(src)
 		if err != nil {
-			return nil, fmt.Errorf("Failed to uncompress gzip file downloaded from %s: %s", url, err)
+			return nil, fmt.Errorf("failed to uncompress gzip file downloaded from %s: %s", url, err)
 		}
 
 		name := r.Header.Name
 		if !matchExecutableName(cmd, name) {
-			return nil, fmt.Errorf("File name '%s' does not match to command '%s' found in %s", name, cmd, url)
+			return nil, fmt.Errorf("file name '%s' does not match to command '%s' found in %s", name, cmd, url)
 		}
 
 		log.Println("Executable file", name, "was found in gzip file")
@@ -115,7 +115,7 @@ func UncompressCommand(src io.Reader, url, cmd string) (io.Reader, error) {
 
 		xzip, err := xz.NewReader(src)
 		if err != nil {
-			return nil, fmt.Errorf("Failed to uncompress .tar.xz file: %s", err)
+			return nil, fmt.Errorf("failed to uncompress .tar.xz file: %s", err)
 		}
 
 		return unarchiveTar(xzip, url, cmd)
@@ -124,7 +124,7 @@ func UncompressCommand(src io.Reader, url, cmd string) (io.Reader, error) {
 
 		xzip, err := xz.NewReader(src)
 		if err != nil {
-			return nil, fmt.Errorf("Failed to uncompress xzip file downloaded from %s: %s", url, err)
+			return nil, fmt.Errorf("failed to uncompress xzip file downloaded from %s: %s", url, err)
 		}
 
 		log.Println("Uncompressed file from xzip is assumed to be an executable", cmd)