+++ selfupdate/update.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -31,7 +30,7 @@ func uncompressAndUpdate(src io.Reader, assetURL, cmdPath string) error {
 func (up *Updater) downloadDirectlyFromURL(assetURL string) (io.ReadCloser, error) {
 	req, err := http.NewRequest("GET", assetURL, nil)
 	if err != nil {
-		return nil, fmt.Errorf("Failed to create HTTP request to %s: %s", assetURL, err)
+		return nil, fmt.Errorf("failed to create HTTP request to %s: %s", assetURL, err)
 	}
 
 	req.Header.Add("Accept", "application/octet-stream")
@@ -42,11 +41,11 @@ func (up *Updater) downloadDirectlyFromURL(assetURL string) (io.ReadCloser, erro
 	// Use default HTTP client instead.
 	res, err := http.DefaultClient.Do(req)
 	if err != nil {
-		return nil, fmt.Errorf("Failed to download a release file from %s: %s", assetURL, err)
+		return nil, fmt.Errorf("failed to download a release file from %s: %s", assetURL, err)
 	}
 
 	if res.StatusCode != 200 {
-		return nil, fmt.Errorf("Failed to download a release file from %s: Not successful status %d", assetURL, res.StatusCode)
+		return nil, fmt.Errorf("failed to download a release file from %s: Not successful status %d", assetURL, res.StatusCode)
 	}
 
 	return res.Body, nil
@@ -59,7 +58,7 @@ func (up *Updater) UpdateTo(rel *Release, cmdPath string) error {
 	var client http.Client
 	src, redirectURL, err := up.api.Repositories.DownloadReleaseAsset(up.apiCtx, rel.RepoOwner, rel.RepoName, rel.AssetID, &client)
 	if err != nil {
-		return fmt.Errorf("Failed to call GitHub Releases API for getting an asset(ID: %d) for repository '%s/%s': %s", rel.AssetID, rel.RepoOwner, rel.RepoName, err)
+		return fmt.Errorf("failed to call GitHub Releases API for getting an asset(ID: %d) for repository '%s/%s': %s", rel.AssetID, rel.RepoOwner, rel.RepoName, err)
 	}
 	if redirectURL != "" {
 		log.Println("Redirect URL was returned while trying to download a release asset from GitHub API. Falling back to downloading from asset URL directly:", redirectURL)
@@ -70,9 +69,9 @@ func (up *Updater) UpdateTo(rel *Release, cmdPath string) error {
 	}
 	defer src.Close()
 
-	data, err := ioutil.ReadAll(src)
+	data, err := io.ReadAll(src)
 	if err != nil {
-		return fmt.Errorf("Failed reading asset body: %v", err)
+		return fmt.Errorf("failed reading asset body: %v", err)
 	}
 
 	if up.validator == nil {
@@ -81,7 +80,7 @@ func (up *Updater) UpdateTo(rel *Release, cmdPath string) error {
 
 	validationSrc, validationRedirectURL, err := up.api.Repositories.DownloadReleaseAsset(up.apiCtx, rel.RepoOwner, rel.RepoName, rel.ValidationAssetID, &client)
 	if err != nil {
-		return fmt.Errorf("Failed to call GitHub Releases API for getting an validation asset(ID: %d) for repository '%s/%s': %s", rel.ValidationAssetID, rel.RepoOwner, rel.RepoName, err)
+		return fmt.Errorf("failed to call GitHub Releases API for getting an validation asset(ID: %d) for repository '%s/%s': %s", rel.ValidationAssetID, rel.RepoOwner, rel.RepoName, err)
 	}
 	if validationRedirectURL != "" {
 		log.Println("Redirect URL was returned while trying to download a release validation asset from GitHub API. Falling back to downloading from asset URL directly:", redirectURL)
@@ -93,13 +92,13 @@ func (up *Updater) UpdateTo(rel *Release, cmdPath string) error {
 
 	defer validationSrc.Close()
 
-	validationData, err := ioutil.ReadAll(validationSrc)
+	validationData, err := io.ReadAll(validationSrc)
 	if err != nil {
-		return fmt.Errorf("Failed reading validation asset body: %v", err)
+		return fmt.Errorf("failed reading validation asset body: %v", err)
 	}
 
 	if err := up.validator.Validate(data, validationData); err != nil {
-		return fmt.Errorf("Failed validating asset content: %v", err)
+		return fmt.Errorf("failed validating asset content: %v", err)
 	}
 
 	return uncompressAndUpdate(bytes.NewReader(data), rel.AssetURL, cmdPath)
@@ -115,12 +114,12 @@ func (up *Updater) UpdateCommand(cmdPath string, current semver.Version, slug st
 
 	stat, err := os.Lstat(cmdPath)
 	if err != nil {
-		return nil, fmt.Errorf("Failed to stat '%s'. File may not exist: %s", cmdPath, err)
+		return nil, fmt.Errorf("failed to stat '%s'. File may not exist: %s", cmdPath, err)
 	}
 	if stat.Mode()&os.ModeSymlink != 0 {
 		p, err := filepath.EvalSymlinks(cmdPath)
 		if err != nil {
-			return nil, fmt.Errorf("Failed to resolve symlink '%s' for executable: %s", cmdPath, err)
+			return nil, fmt.Errorf("failed to resolve symlink '%s' for executable: %s", cmdPath, err)
 		}
 		cmdPath = p
 	}