+++ selfupdate/update_test.go
@@ -148,7 +148,7 @@ func TestUpdateBrokenSymlinks(t *testing.T) {
 		if err == nil {
 			t.Fatal("Error should occur for unlinked symlink", p)
 		}
-		if !strings.Contains(err.Error(), "Failed to resolve symlink") {
+		if !strings.Contains(err.Error(), "failed to resolve symlink") {
 			t.Fatal("Unexpected error for broken symlink", p, err)
 		}
 	}
@@ -221,7 +221,7 @@ func TestBrokenBinaryUpdate(t *testing.T) {
 	if err == nil {
 		t.Fatal("Error should occur for broken package")
 	}
-	if !strings.Contains(err.Error(), "Failed to uncompress .tar.gz file") {
+	if !strings.Contains(err.Error(), "failed to uncompress .tar.gz file") {
 		t.Fatal("Unexpected error:", err)
 	}
 }
@@ -232,7 +232,7 @@ func TestInvalidSlugForUpdate(t *testing.T) {
 	if err == nil {
 		t.Fatal("Unknown repo should cause an error")
 	}
-	if !strings.Contains(err.Error(), "Invalid slug format") {
+	if !strings.Contains(err.Error(), "invalid slug format") {
 		t.Fatal("Unexpected error:", err)
 	}
 }
@@ -242,7 +242,7 @@ func TestInvalidAssetURL(t *testing.T) {
 	if err == nil {
 		t.Fatal("Error should occur for URL not found")
 	}
-	if !strings.Contains(err.Error(), "Failed to download a release file") {
+	if !strings.Contains(err.Error(), "failed to download a release file") {
 		t.Fatal("Unexpected error:", err)
 	}
 }
@@ -253,101 +253,7 @@ func TestBrokenAsset(t *testing.T) {
 	if err == nil {
 		t.Fatal("Error should occur for URL not found")
 	}
-	if !strings.Contains(err.Error(), "Failed to uncompress zip file") {
+	if !strings.Contains(err.Error(), "failed to uncompress zip file") {
 		t.Fatal("Unexpected error:", err)
-	}
-}
-
-func TestBrokenGitHubEnterpriseURL(t *testing.T) {
-	up, err := NewUpdater(Config{APIToken: "hogehoge", EnterpriseBaseURL: "https://example.com"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = up.UpdateTo(&Release{AssetURL: "https://example.com"}, "foo")
-	if err == nil {
-		t.Fatal("Invalid GitHub Enterprise base URL should raise an error")
-	}
-	if !strings.Contains(err.Error(), "Failed to call GitHub Releases API for getting an asset") {
-		t.Error("Unexpected error occurred:", err)
-	}
-}
-
-func TestUpdateFromGitHubEnterprise(t *testing.T) {
-	token := os.Getenv("GITHUB_ENTERPRISE_TOKEN")
-	base := os.Getenv("GITHUB_ENTERPRISE_BASE_URL")
-	repo := os.Getenv("GITHUB_ENTERPRISE_REPO")
-	if token == "" {
-		t.Skip("because token for GHE is not found")
-	}
-	if base == "" {
-		t.Skip("because base URL for GHE is not found")
-	}
-	if repo == "" {
-		t.Skip("because repo slug for GHE is not found")
-	}
-
-	setupTestBinary()
-	defer teardownTestBinary()
-
-	up, err := NewUpdater(Config{APIToken: token, EnterpriseBaseURL: base})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	latest := semver.MustParse("1.2.3")
-	prev := semver.MustParse("1.2.2")
-	rel, err := up.UpdateCommand("github-release-test", prev, repo)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if rel.Version.NE(latest) {
-		t.Error("Version is not latest", rel.Version)
-	}
-
-	bytes, err := exec.Command(filepath.FromSlash("./github-release-test")).Output()
-	if err != nil {
-		t.Fatal("Failed to run test binary after update:", err)
-	}
-
-	out := string(bytes)
-	if out != "v1.2.3\n" {
-		t.Error("Output from test binary after update is unexpected:", out)
-	}
-}
-
-func TestUpdateFromGitHubPrivateRepo(t *testing.T) {
-	token := os.Getenv("GITHUB_PRIVATE_TOKEN")
-	if token == "" {
-		t.Skip("because GITHUB_PRIVATE_TOKEN is not set")
-	}
-
-	setupTestBinary()
-	defer teardownTestBinary()
-
-	up, err := NewUpdater(Config{APIToken: token})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	latest := semver.MustParse("1.2.3")
-	prev := semver.MustParse("1.2.2")
-	rel, err := up.UpdateCommand("github-release-test", prev, "rhysd/private-release-test")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if rel.Version.NE(latest) {
-		t.Error("Version is not latest", rel.Version)
-	}
-
-	bytes, err := exec.Command(filepath.FromSlash("./github-release-test")).Output()
-	if err != nil {
-		t.Fatal("Failed to run test binary after update:", err)
-	}
-
-	out := string(bytes)
-	if out != "v1.2.3\n" {
-		t.Error("Output from test binary after update is unexpected:", out)
 	}
 }